requirements.txt
asgiref==3.5.2
bcrypt==3.2.2
black==22.3.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cloudinary==1.29.0
commonmark==0.9.1
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.4
django-cloudinary-storage==0.3.0
django-cors-headers==3.12.0
django-multiselectfield==0.1.12
django-oauth-toolkit==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
httpie==3.2.1
idna==3.3
jwcrypto==1.3.1
multidict==6.0.2
mypy-extensions==0.4.3
oauthlib==3.2.0
passlib==1.7.4
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
PySocks==1.7.1
python-decouple==3.6
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.9.1
rich==12.4.4
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.9
uuid==1.30
whitenoise==6.2.0
wrapt==1.14.1
